+++ assets/script.js
@@ -1,19 +1,39 @@
 /* eslint-env browser */
-/* global Prism */
+/* global examples, Prism */
 
 document.addEventListener('DOMContentLoaded', function () {
-	Array.prototype.forEach.call(document.querySelectorAll('pre code'), function (code) {
-		var match = code.className.match(/(?:^|\s)lang-(\w+?)(_example|\s|$)/) || [];
-		var lang  = match[1];
-		var demo  = !!match[2];
-		var text  = String(code.textContent || code.innerText || '');
+	Array.prototype.forEach.call(document.querySelectorAll('pre code[class^="lang"]'), function (code) {
+		// set pre, wrap, opts, and get meta data from code
+		var pre  = code.parentNode;
+		var wrap = pre.parentNode.insertBefore(document.createElement('figure'), pre);
+		var conf = {};
+		var text = String(code.textContent || code.innerText || '');
 
-		if (lang && lang in Prism.languages) {
-			code.innerHTML = Prism.highlight(text, Prism.languages[lang]);
-		}
+		// get meta data from code class
+		code.className.replace(/^lang-(\w+)(?::(\w+))?/, function ($0, $1, $2) {
+			if ($2) return 'example:' + $2 + ',lang:' + $2;
 
-		if (demo && lang in examples) {
-			examples[lang](text, code);
+			if ($1 === 'example') return 'example:html';
+
+			return 'lang:' + $1;
+		}).split(/\s*,\s*/).forEach(function (opt) {
+			opt = opt.split(':');
+
+			conf[opt.shift().trim()] = opt.join(':').trim();
+		});
+
+		code.removeAttribute('class');
+
+		wrap.appendChild(pre);
+
+		// conditionally syntax highlight code
+		if (conf.lang in Prism.languages) code.innerHTML = Prism.highlight(text, Prism.languages[conf.lang]);
+
+		// conditionally create code examples
+		if (conf.example in examples.lang) {
+			examples.lang[conf.example](pre, text, conf);
+
+			if (!(conf.lang in Prism.languages)) wrap.removeChild(pre);
 		}
 	});
 });